refactor(example): move src/ to top-level
