+++ example/vite-demo/index.html
@@ -10,8 +10,8 @@
       <a href='https://github.com/gsuuon/ad-llama'><span id='gh'></span></a>
       ad-llama
     </h1>
-    <div><a href="https://github.com/gsuuon/ad-llama/blob/main/example/vite-demo/src/main.ts">code</a></div>
+    <div><a href="https://github.com/gsuuon/ad-llama/blob/main/example/vite-demo/main.ts">code</a></div>
     <div id="app"></div>
-    <script type="module" src="/src/main.ts"></script>
+    <script type="module" src="/main.ts"></script>
   </body>
 </html>
