+++ example/vite-demo/main.ts
