+++ example/vite-demo/renderTemplate.ts
