+++ example/vite-demo/tsconfig.json
@@ -19,5 +19,6 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src", "hn"]
+  "include": ["."],
+  "exclude": ["node_modules", "dist", "public", ".vercel"]
 }
