+++ example/vite-demo/vite-env.d.ts
